Allowed recoil configs to have decimals
There really was no need to have recoil be an integer. It was being added to a float anyways.

+++ src/main/java/com/flansmod/common/guns/GunType.java
@@ -39,7 +39,7 @@ public class GunType extends PaintableType implements IScope
 	/**
 	 * The amount to recoil the player's view by when firing a single shot from this gun
 	 */
-	public int recoil;
+	public float recoil;
 	/**
 	 * The amount that bullets spread out when fired from this gun
 	 */
